add exercises

+++ C语言程序设计（乐学）/计算子字符串个数/1.c
@@ -16,7 +16,7 @@ main()
 
 #include <string.h>
 int countsub(char* str, char* ss) {
-    size_t len1 = strlen(str), len2 = strlen(ss);
+    unsigned len1 = strlen(str), len2 = strlen(ss);
     int ptr = 0, cnt = 0, ans = 0;
     while (ptr < len1) {
         if (strstr(str + ptr, ss) == str + ptr) {
